+++ photoprism/CHANGELOG.md
@@ -1,4 +1,7 @@
 
+## ubuntu-2026-03-05-1 (2026-05-10)
+- Fix MariaDB connection on HAOS >=17.3 by forcing IPv4 host resolution (#2688)
+
 ## ubuntu-2026-03-05 (2026-03-05)
 - Update to latest version from photoprism/photoprism
 ## ubuntu-2025-11-30-4 (26-02-2026)