+++ zoneminder/CHANGELOG.md
@@ -1,4 +1,7 @@
 
+## 1.38.1-1 (2026-05-10)
+- Fix MariaDB connection on HAOS >=17.3 by forcing IPv4 host resolution (#2688)
+
 ## 1.38.1 (2026-02-21)
 - Update to latest version from zoneminder-containers/zoneminder-base (changelog : https://github.com/zoneminder-containers/zoneminder-base/releases)